Moved the JSON adapters into one adapter package
